refactor(homepage): move language dropdown menu component off from navigator component
