+++ homepage/src/layouts/header/navigator.jsx
@@ -1,30 +1,6 @@
-import { useRouter } from 'next/router';
 import Image from 'next/image';
 import Link from 'next/link';
-
-const languageLabelDictionary = {
-  en: 'English',
-  'zh-tw': '中文',
-};
-
-const LanguageDropdownMenu = () => {
-  const router = useRouter();
-  const { asPath, locales } = router;
-  return (
-    <div className="dropdown-menu">
-      {locales.map((locale) => (
-        <Link
-          key={languageLabelDictionary[locale]}
-          className="dropdown-item"
-          href={`/${locale}${asPath}`}
-          locale={false}
-        >
-          {languageLabelDictionary[locale]}
-        </Link>
-      ))}
-    </div>
-  );
-};
+import LanguageDropdownMenu from './languageDropdownMenu';
 
 const Navigator = ({ logo, navigationItems }) => (
   <nav className="navbar navbar-expand-lg">
